[Kinetics] finalize ReactionRate and corresponding data types
